+++ game/content/monster/Sahagin.lua
@@ -11,7 +11,7 @@ local FlinchAnimation = require("arpg/field/animation/FlinchAnimation");
 local IdleAnimation = require("arpg/field/animation/IdleAnimation");
 local WalkAnimation = require("arpg/field/animation/WalkAnimation");
 local CommonShader = require("arpg/graphics/CommonShader");
-local MovementAI = require("engine/mapscene/behavior/ai/movement/MovementAI");
+local MovementAI = require("engine/mapscene/behavior/ai/MovementAI");
 local Entity = require("engine/ecs/Entity");
 local Assets = require("engine/resources/Assets");
 local Actor = require("engine/mapscene/behavior/Actor");
