+++ game/engine/TestSuite.lua
@@ -1,11 +1,11 @@
 local Module = require("engine/Module");
 
 local engineTestFiles = {
-	"engine/mapscene/behavior/ai/movement/TestAlignGoal",
-	"engine/mapscene/behavior/ai/movement/TestEntityGoal",
-	"engine/mapscene/behavior/ai/movement/TestMovement",
-	"engine/mapscene/behavior/ai/navmesh/TestPath",
-	"engine/mapscene/behavior/ai/movement/TestPositionGoal",
+	"engine/mapscene/behavior/ai/TestAlignGoal",
+	"engine/mapscene/behavior/ai/TestEntityGoal",
+	"engine/mapscene/behavior/ai/TestMovement",
+	"engine/mapscene/behavior/ai/TestPath",
+	"engine/mapscene/behavior/ai/TestPositionGoal",
 	"engine/dev/cli/TestCLI",
 	"engine/ecs/TestECS",
 	"engine/input/TestInputDevice",
