+++ game/engine/mapscene/MapScene.lua
@@ -5,7 +5,7 @@ local ECS = require("engine/ecs/ECS");
 local Camera = require("engine/mapscene/Camera");
 local Scene = require("engine/Scene");
 local ActorSystem = require("engine/mapscene/behavior/ActorSystem");
-local MovementAISystem = require("engine/mapscene/behavior/ai/movement/MovementAISystem");
+local MovementAISystem = require("engine/mapscene/behavior/ai/MovementAISystem");
 local InputListenerSystem = require("engine/mapscene/behavior/InputListenerSystem");
 local ScriptRunnerSystem = require("engine/mapscene/behavior/ScriptRunnerSystem");
 local SpriteSystem = require("engine/mapscene/display/SpriteSystem");
