+++ game/engine/mapscene/behavior/ai/NavigationFailureEvent.lua
