+++ game/engine/resources/map/TestNavigationMesh.lua
@@ -10,19 +10,19 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Load Navmesh Lua file"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- end
 
 -- tests[#tests + 1] = {name = "Generate navmesh for empty map"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, 0);
 -- end
 
 -- tests[#tests + 1] = {name = "Generate navmesh for empty map with padding"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local padding = 1;
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, padding);
@@ -30,7 +30,7 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Generate navmesh for empty map with extreme padding"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local padding = 20;
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, padding);
@@ -38,7 +38,7 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Find path in empty map"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, 0);
 -- 	local path = navmesh:findPath(1, 2, 8, 9);
@@ -51,7 +51,7 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Find path from outside navmesh"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, 0);
 -- 	local path = navmesh:findPath(-4, 2, 8, 9);
@@ -65,7 +65,7 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Find path to outside navmesh"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, 0);
 -- 	local path = navmesh:findPath(3, 5, 8, 14);
@@ -78,7 +78,7 @@ local tests = {};
 
 -- tests[#tests + 1] = {name = "Project point on navmesh"};
 -- tests[#tests].body = function()
--- 	local Navmesh = require("engine/mapscene/behavior/ai/navmesh/Navmesh");
+-- 	local Navmesh = require("engine/mapscene/behavior/ai/Navmesh");
 -- 	local collisionMesh = CollisionMesh:new(10, 10, 5, 5);
 -- 	local navmesh = Navmesh:new(10, 10, collisionMesh, 0);
 -- 	local px, py = navmesh:getNearestPointOnNavmesh(-5, -5);
